Export an object with `init` method for `click` and `view`.
So `oTracking.click` becomes `oTracking.click.init` and
`oTracking.view` becomes `oTracking.view.init`.

- The documentation refers to `oTracking.click.init` and
`oTracking.view.init`. `oTracking.click` and `oTracking.view` are
undocumented.
- `oTracking.page.init` already exists, each event type will align.

This could be considered a breaking change but I think we should
treat this as a bugfix since there are few users of v3 and the
methods being changed are not documented.