Add failing test: new context should clear queue of parent context
When a link is clicked to the same domain the event is not fired,
instead it is added to a click queue which is stored in local
storage. When the next page loads click events are sent from the
queue. If the user clicks a link to open it in a new tab (cmd+click
on macOS) the original, parent context maintains a copy of the
queue in memory and re-inserts it into local storage after the
events have already been sent. This causes duplicate events
(with differeing context ids) to be sent when another link is
clicked on the parent context.